pyproject.toml
[tool.poetry]
name = "metadata"
version = "2.2.5"
description = "Create and update file metadata"
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.88.0"
fastapi-health = "0.4.0"
fastapi-sqlalchemy = "0.2.1"
fastapi-utils = "0.2.1"
psycopg2-binary = "2.9.3"
requests = "2.24.0"
sqlalchemy-utils = "0.38.2"
uvicorn = "0.17.5"
python-json-logger = "2.0.2"
fastavro = "1.6.0"
kafka-python = "^2.0.2"
pilot-platform-common = "0.3.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-httpx = "0.21.0"
pytest-cov = "3.0.0"
flake8 = "4.0.1"
alembic = "^1.7.7"
faker = "14.2.0"
pytest-mock = "^3.10.0"
testcontainers = "3.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
